Move ReleaseResolution into Version
